Move all classes in IDrawable.java to their own class file
