+++ src/games/strategy/triplea/ui/screen/drawable/IDrawable.java
@@ -71,723 +71,3 @@ public interface IDrawable {
 }
 
 
-class DrawableComparator implements Comparator<IDrawable> {
-  @Override
-  public int compare(final IDrawable o1, final IDrawable o2) {
-    return o1.getLevel() - o2.getLevel();
-  }
-}
-
-
-class TerritoryNameDrawable implements IDrawable {
-  private final String m_territoryName;
-  private final IUIContext m_uiContext;
-
-  public TerritoryNameDrawable(final String territoryName, final IUIContext context) {
-    this.m_territoryName = territoryName;
-    this.m_uiContext = context;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Territory territory = data.getMap().getTerritory(m_territoryName);
-    final TerritoryAttachment ta = TerritoryAttachment.get(territory);
-    final boolean drawFromTopLeft = mapData.drawNamesFromTopLeft();
-    final boolean showSeaNames = mapData.drawSeaZoneNames();
-    final boolean showComments = mapData.drawComments();
-    boolean drawComments = false;
-    String commentText = null;
-    if (territory.isWater()) {
-      // this is for special comments, like convoy zones, etc.
-      if (ta != null && showComments) {
-        if (ta.getConvoyRoute() && ta.getProduction() > 0 && ta.getOriginalOwner() != null) {
-          drawComments = true;
-          if (ta.getConvoyAttached().isEmpty()) {
-            commentText = MyFormatter
-                .defaultNamedToTextList(TerritoryAttachment.getWhatTerritoriesThisIsUsedInConvoysFor(territory, data))
-                + " " + ta.getOriginalOwner().getName() + " Blockade Route";
-          } else {
-            commentText = MyFormatter.defaultNamedToTextList(ta.getConvoyAttached()) + " "
-                + ta.getOriginalOwner().getName() + " Convoy Route";
-          }
-        } else if (ta.getConvoyRoute()) {
-          drawComments = true;
-          if (ta.getConvoyAttached().isEmpty()) {
-            commentText = MyFormatter.defaultNamedToTextList(
-                TerritoryAttachment.getWhatTerritoriesThisIsUsedInConvoysFor(territory, data)) + " Blockade Route";
-          } else {
-            commentText = MyFormatter.defaultNamedToTextList(ta.getConvoyAttached()) + " Convoy Route";
-          }
-        } else if (ta.getProduction() > 0 && ta.getOriginalOwner() != null) {
-          drawComments = true;
-          final PlayerID originalOwner = ta.getOriginalOwner();
-          commentText = originalOwner.getName() + " Convoy Center";
-        }
-      }
-      if (!drawComments && !showSeaNames) {
-        return;
-      }
-    }
-
-    graphics.setFont(MapImage.getPropertyMapFont());
-    graphics.setColor(MapImage.getPropertyTerritoryNameAndPUAndCommentcolor());
-    final FontMetrics fm = graphics.getFontMetrics();
-
-    // if we specify a placement point, use it otherwise try to center it
-    int x;
-    int y;
-    final Optional<Point> namePlace = mapData.getNamePlacementPoint(territory);
-    if (namePlace.isPresent()) {
-      x = namePlace.get().x;
-      y = namePlace.get().y;
-    } else {
-      final Rectangle territoryBounds = getBestTerritoryNameRect(mapData, territory, fm);
-      x = territoryBounds.x + (int) territoryBounds.getWidth() / 2 - fm.stringWidth(territory.getName()) / 2;
-      y = territoryBounds.y + (int) territoryBounds.getHeight() / 2 + fm.getAscent() / 2;
-    }
-
-    // draw comments above names
-    if (showComments && drawComments && commentText != null) {
-      final Optional<Point> place = mapData.getCommentMarkerLocation(territory);
-      if (place.isPresent()) {
-        draw(bounds, graphics, place.get().x, place.get().y, null, commentText, drawFromTopLeft);
-      } else {
-        draw(bounds, graphics, x, y - fm.getHeight(), null, commentText, drawFromTopLeft);
-      }
-    }
-    // draw territory names
-    if (mapData.drawTerritoryNames() && mapData.shouldDrawTerritoryName(m_territoryName)) {
-      if (!territory.isWater() || showSeaNames) {
-        final Image nameImage = mapData.getTerritoryNameImages().get(territory.getName());
-        draw(bounds, graphics, x, y, nameImage, territory.getName(), drawFromTopLeft);
-      }
-    }
-    // draw the PUs.
-    if (ta != null && ta.getProduction() > 0 && mapData.drawResources()) {
-      final Image img = m_uiContext.getPUImageFactory().getPUImage(ta.getProduction());
-      final String prod = Integer.valueOf(ta.getProduction()).toString();
-      final Optional<Point> place = mapData.getPUPlacementPoint(territory);
-      // if pu_place.txt is specified draw there
-      if (place.isPresent()) {
-        draw(bounds, graphics, place.get().x, place.get().y, img, prod, drawFromTopLeft);
-      } else {
-        // otherwise, draw under the territory name
-        draw(bounds, graphics, x + ((fm.stringWidth(m_territoryName)) >> 1) - ((fm.stringWidth(prod)) >> 1),
-            y + fm.getLeading() + fm.getAscent(), img, prod, drawFromTopLeft);
-      }
-    }
-  }
-
-  /**
-   * Find the best rectangle inside the territory to place the name in. Finds the rectangle
-   * that can fit the name, that is the closest to the vertical center, and has a large width at
-   * that location. If there isn't any rectangles that can fit the name then default back to the
-   * bounding rectangle.
-   */
-  private Rectangle getBestTerritoryNameRect(final MapData mapData, final Territory territory,
-      final FontMetrics fontMetrics) {
-
-    // Find bounding rectangle and parameters for creating a grid (20 x 20) across the territory
-    final Rectangle territoryBounds = mapData.getBoundingRect(territory);
-    Rectangle result = territoryBounds;
-    final int maxX = territoryBounds.x + territoryBounds.width;
-    final int maxY = territoryBounds.y + territoryBounds.height;
-    final int centerY = territoryBounds.y + territoryBounds.height / 2;
-    final int incrementX = (int) Math.ceil(territoryBounds.width / 20.0);
-    final int incrementY = (int) Math.ceil(territoryBounds.height / 20.0);
-    final int nameWidth = fontMetrics.stringWidth(territory.getName());
-    final int nameHeight = fontMetrics.getAscent();
-    int maxScore = 0;
-
-    // Loop through the grid moving the starting point and determining max width at that point
-    for (int x = territoryBounds.x; x < maxX - nameWidth; x += incrementX) {
-      for (int y = territoryBounds.y; y < maxY - nameHeight; y += incrementY) {
-        for (int endX = maxX; endX > x; endX -= incrementX) {
-          final Rectangle rectangle = new Rectangle(x, y, endX - x, nameHeight);
-
-          // Ranges from 0 when at very top or bottom of territory to height/2 when at vertical center
-          final int verticalDistanceFromEdge = territoryBounds.height / 2 - Math.abs(centerY - nameHeight - y);
-
-          // Score rectangle based on how close to vertical center and territory width at location
-          final int score = verticalDistanceFromEdge * rectangle.width;
-          if (rectangle.width > nameWidth && score > maxScore) {
-
-            // Check to make sure rectangle is contained in the territory
-            if (isRectangleContainedInTerritory(rectangle, territory, mapData)) {
-              maxScore = score;
-              result = rectangle;
-              break;
-            }
-          }
-        }
-      }
-    }
-    return result;
-  }
-
-  private boolean isRectangleContainedInTerritory(final Rectangle rectangle, final Territory territory,
-      final MapData mapData) {
-    final List<Polygon> polygons = mapData.getPolygons(territory.getName());
-    for (final Polygon polygon : polygons) {
-      if (polygon.contains(rectangle)) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  private void draw(final Rectangle bounds, final Graphics2D graphics, final int x, final int y, final Image img,
-      final String prod, final boolean drawFromTopLeft) {
-    int yNormal = y;
-    if (img == null) {
-      if (graphics.getFont().getSize() <= 0) {
-        return;
-      }
-      if (drawFromTopLeft) {
-        final FontMetrics fm = graphics.getFontMetrics();
-        yNormal += fm.getHeight();
-      }
-      graphics.drawString(prod, x - bounds.x, yNormal - bounds.y);
-    } else {
-      // we want to be consistent
-      // drawString takes y as the base line position
-      // drawImage takes x as the top right corner
-      if (!drawFromTopLeft) {
-        yNormal -= img.getHeight(null);
-      }
-      graphics.drawImage(img, x - bounds.x, yNormal - bounds.y, null);
-    }
-  }
-
-  @Override
-  public int getLevel() {
-    return TERRITORY_TEXT_LEVEL;
-  }
-}
-
-
-class VCDrawable implements IDrawable {
-  private final Territory m_location;
-
-  public VCDrawable(final Territory location) {
-    m_location = location;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Point point = mapData.getVCPlacementPoint(m_location);
-    graphics.drawImage(mapData.getVCImage(), point.x - bounds.x, point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return VC_MARKER_LEVEL;
-  }
-}
-
-
-class DecoratorDrawable implements IDrawable {
-  private final Point m_point;
-  private final Image m_image;
-
-  public DecoratorDrawable(final Point point, final Image image) {
-    super();
-    m_point = point;
-    m_image = image;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    graphics.drawImage(m_image, m_point.x - bounds.x, m_point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return DECORATOR_LEVEL;
-  }
-}
-
-
-class CapitolMarkerDrawable implements IDrawable {
-  private final String m_player;
-  private final String m_location;
-  private final IUIContext m_uiContext;
-
-  public CapitolMarkerDrawable(final PlayerID player, final Territory location, final IUIContext uiContext) {
-    super();
-    if (player == null) {
-      throw new IllegalStateException("no player for capitol:" + location);
-    }
-    m_player = player.getName();
-    m_location = location.getName();
-    m_uiContext = uiContext;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    // Changed back to use Large flags
-    final Image img = m_uiContext.getFlagImageFactory().getLargeFlag(data.getPlayerList().getPlayerID(m_player));
-    final Point point = mapData.getCapitolMarkerLocation(data.getMap().getTerritory(m_location));
-    graphics.drawImage(img, point.x - bounds.x, point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return CAPITOL_MARKER_LEVEL;
-  }
-}
-
-
-abstract class MapTileDrawable implements IDrawable {
-  protected boolean m_noImage = false;
-  protected final int m_x;
-  protected final int m_y;
-  protected final IUIContext m_uiContext;
-  protected boolean m_unscaled;
-
-  public MapTileDrawable(final int x, final int y, final IUIContext uiContext) {
-    m_x = x;
-    m_y = y;
-    m_uiContext = uiContext;
-    m_unscaled = false;
-  }
-
-  public abstract MapTileDrawable getUnscaledCopy();
-
-  protected abstract Image getImage();
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Image img = getImage();
-    if (img == null) {
-      return;
-    }
-    final Object oldRenderingValue = graphics.getRenderingHint(RenderingHints.KEY_RENDERING);
-    final Object oldAlphaValue = graphics.getRenderingHint(RenderingHints.KEY_ALPHA_INTERPOLATION);
-    final Object oldInterpolationValue = graphics.getRenderingHint(RenderingHints.KEY_INTERPOLATION);
-    graphics.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_SPEED);
-    graphics.setRenderingHint(RenderingHints.KEY_ALPHA_INTERPOLATION, RenderingHints.VALUE_ALPHA_INTERPOLATION_SPEED);
-    graphics.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_NEAREST_NEIGHBOR);
-    // the tile images are already scaled
-    if (unscaled != null) {
-      graphics.setTransform(unscaled);
-    }
-    final Stopwatch drawStopWatch = new Stopwatch(s_logger, Level.FINEST, "drawing tile images");
-    graphics.drawImage(img, m_x * TileManager.TILE_SIZE - bounds.x, m_y * TileManager.TILE_SIZE - bounds.y, null);
-    drawStopWatch.done();
-    if (unscaled != null) {
-      graphics.setTransform(scaled);
-    }
-    if (oldAlphaValue == null) {
-      graphics.setRenderingHint(RenderingHints.KEY_ALPHA_INTERPOLATION,
-          RenderingHints.VALUE_ALPHA_INTERPOLATION_DEFAULT);
-    } else {
-      graphics.setRenderingHint(RenderingHints.KEY_ALPHA_INTERPOLATION, oldAlphaValue);
-    }
-    if (oldRenderingValue == null) {
-      graphics.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_DEFAULT);
-    } else {
-      graphics.setRenderingHint(RenderingHints.KEY_RENDERING, oldRenderingValue);
-    }
-    if (oldInterpolationValue == null) {
-      graphics.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_NEAREST_NEIGHBOR);
-    } else {
-      graphics.setRenderingHint(RenderingHints.KEY_INTERPOLATION, oldInterpolationValue);
-    }
-  }
-}
-
-
-class ReliefMapDrawable extends MapTileDrawable {
-  public ReliefMapDrawable(final int x, final int y, final IUIContext context) {
-    super(x, y, context);
-  }
-
-  @Override
-  public MapTileDrawable getUnscaledCopy() {
-    final ReliefMapDrawable copy = new ReliefMapDrawable(m_x, m_y, m_uiContext);
-    copy.m_unscaled = true;
-    return copy;
-  }
-
-  @Override
-  protected Image getImage() {
-    if (m_noImage) {
-      return null;
-    }
-    if (!TileImageFactory.getShowReliefImages()) {
-      return null;
-    }
-    Image rVal;
-    if (m_unscaled) {
-      rVal = m_uiContext.getTileImageFactory().getUnscaledUncachedReliefTile(m_x, m_y);
-    } else {
-      rVal = m_uiContext.getTileImageFactory().getReliefTile(m_x, m_y);
-    }
-    if (rVal == null) {
-      m_noImage = true;
-    }
-    return rVal;
-  }
-
-  @Override
-  public int getLevel() {
-    return RELIEF_LEVEL;
-  }
-}
-
-
-class BaseMapDrawable extends MapTileDrawable {
-  public BaseMapDrawable(final int x, final int y, final IUIContext uiContext) {
-    super(x, y, uiContext);
-  }
-
-  @Override
-  public MapTileDrawable getUnscaledCopy() {
-    final BaseMapDrawable copy = new BaseMapDrawable(m_x, m_y, m_uiContext);
-    copy.m_unscaled = true;
-    return copy;
-  }
-
-  @Override
-  protected Image getImage() {
-    if (m_noImage) {
-      return null;
-    }
-    Image rVal;
-    if (m_unscaled) {
-      rVal = m_uiContext.getTileImageFactory().getUnscaledUncachedBaseTile(m_x, m_y);
-    } else {
-      rVal = m_uiContext.getTileImageFactory().getBaseTile(m_x, m_y);
-    }
-    if (rVal == null) {
-      m_noImage = true;
-    }
-    return rVal;
-  }
-
-  @Override
-  public int getLevel() {
-    return BASE_MAP_LEVEL;
-  }
-}
-
-
-// Rewritten class to use country markers rather than shading for Convoy Centers/Routes.
-class ConvoyZoneDrawable implements IDrawable {
-  private final String m_player;
-  private final String m_location;
-  private final IUIContext m_uiContext;
-
-  public ConvoyZoneDrawable(final PlayerID player, final Territory location, final IUIContext uiContext) {
-    super();
-    m_player = player.getName();
-    m_location = location.getName();
-    m_uiContext = uiContext;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    Image img;
-    if (mapData.useNation_convoyFlags()) {
-      img = m_uiContext.getFlagImageFactory().getConvoyFlag(data.getPlayerList().getPlayerID(m_player));
-    } else {
-      img = m_uiContext.getFlagImageFactory().getFlag(data.getPlayerList().getPlayerID(m_player));
-    }
-    final Point point = mapData.getConvoyMarkerLocation(data.getMap().getTerritory(m_location));
-    graphics.drawImage(img, point.x - bounds.x, point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return CAPITOL_MARKER_LEVEL;
-  }
-}
-
-
-// Class to use 'Faded' country markers for Kamikaze Zones.
-class KamikazeZoneDrawable implements IDrawable {
-  private final String m_location;
-  private final IUIContext m_uiContext;
-
-  public KamikazeZoneDrawable(final Territory location, final IUIContext uiContext2) {
-    super();
-    m_location = location.getName();
-    m_uiContext = uiContext2;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    // Change so only original owner gets the kamikazi zone marker
-    final Territory terr = data.getMap().getTerritory(m_location);
-    final TerritoryAttachment ta = TerritoryAttachment.get(terr);
-    PlayerID owner = null;
-    if (games.strategy.triplea.Properties.getKamikazeSuicideAttacksDoneByCurrentTerritoryOwner(data)) {
-      owner = terr.getOwner();
-      if (owner == null) {
-        owner = PlayerID.NULL_PLAYERID;
-      }
-    } else {
-      if (ta == null) {
-        owner = PlayerID.NULL_PLAYERID;
-      } else {
-        owner = ta.getOriginalOwner();
-        if (owner == null) {
-          owner = PlayerID.NULL_PLAYERID;
-        }
-      }
-    }
-    final Image img = m_uiContext.getFlagImageFactory().getFadedFlag(owner);
-    final Point point = mapData.getKamikazeMarkerLocation(data.getMap().getTerritory(m_location));
-    graphics.drawImage(img, point.x - bounds.x, point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return CAPITOL_MARKER_LEVEL;
-  }
-}
-
-
-class BlockadeZoneDrawable implements IDrawable {
-  private final String m_location;
-
-  // private final UIContext m_uiContext;
-  public BlockadeZoneDrawable(final Territory location, final IUIContext uiContext) {
-    super();
-    m_location = location.getName();
-    // m_uiContext = uiContext;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    // Find blockade.png from misc folder
-    final Point point = mapData.getBlockadePlacementPoint(data.getMap().getTerritory(m_location));
-    graphics.drawImage(mapData.getBlockadeImage(), point.x - bounds.x, point.y - bounds.y, null);
-  }
-
-  @Override
-  public int getLevel() {
-    return CAPITOL_MARKER_LEVEL;
-  }
-}
-
-
-class TerritoryEffectDrawable implements IDrawable {
-  private final TerritoryEffect m_effect;
-  private final Point m_point;
-
-  public TerritoryEffectDrawable(final TerritoryEffect te, final Point point) {
-    super();
-    m_effect = te;
-    m_point = point;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    graphics.drawImage(mapData.getTerritoryEffectImage(m_effect.getName()), m_point.x - bounds.x, m_point.y - bounds.y,
-        null);
-  }
-
-  @Override
-  public int getLevel() {
-    return TERRITORY_EFFECT_LEVEL;
-  }
-}
-
-
-class OptionalExtraTerritoryBordersDrawable implements IDrawable {
-  private final String m_territoryName;
-  private final OptionalExtraBorderLevel m_level;
-
-  public OptionalExtraTerritoryBordersDrawable(final String territoryName, final OptionalExtraBorderLevel level) {
-    m_territoryName = territoryName;
-    m_level = level;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Territory territory = data.getMap().getTerritory(m_territoryName);
-    final List<Polygon> polys = mapData.getPolygons(territory);
-    final Iterator<Polygon> iter2 = polys.iterator();
-    while (iter2.hasNext()) {
-      Polygon polygon = iter2.next();
-      // if we dont have to draw, dont
-      if (!polygon.intersects(bounds) && !polygon.contains(bounds)) {
-        continue;
-      }
-      // use a copy since we will move the polygon
-      polygon = new Polygon(polygon.xpoints, polygon.ypoints, polygon.npoints);
-      polygon.translate(-bounds.x, -bounds.y);
-      graphics.setColor(Color.BLACK);
-      graphics.drawPolygon(polygon);
-    }
-  }
-
-  @Override
-  public int getLevel() {
-    if (m_level == OptionalExtraBorderLevel.HIGH) {
-      return OPTIONAL_EXTRA_TERRITORY_BORDERS_HIGH_LEVEL;
-    }
-    return OPTIONAL_EXTRA_TERRITORY_BORDERS_MEDIUM_LEVEL;
-  }
-}
-
-
-class SeaZoneOutlineDrawable implements IDrawable {
-  private final String m_territoryName;
-
-  public SeaZoneOutlineDrawable(final String territoryName) {
-    m_territoryName = territoryName;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Territory territory = data.getMap().getTerritory(m_territoryName);
-    final List<Polygon> polys = mapData.getPolygons(territory);
-    final Iterator<Polygon> iter2 = polys.iterator();
-    while (iter2.hasNext()) {
-      Polygon polygon = iter2.next();
-      // if we dont have to draw, dont
-      if (!polygon.intersects(bounds) && !polygon.contains(bounds)) {
-        continue;
-      }
-      // use a copy since we will move the polygon
-      polygon = new Polygon(polygon.xpoints, polygon.ypoints, polygon.npoints);
-      polygon.translate(-bounds.x, -bounds.y);
-      graphics.setColor(Color.BLACK);
-      graphics.drawPolygon(polygon);
-    }
-  }
-
-  @Override
-  public int getLevel() {
-    return POLYGONS_LEVEL;
-  }
-}
-
-
-abstract class TerritoryDrawable {
-  protected static void draw(final Rectangle bounds, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform scaled, final Territory territory,
-      final Paint territoryPaint) {
-    final List<Polygon> polys = mapData.getPolygons(territory);
-    for (Polygon polygon : polys) {
-      // if we dont have to draw, dont
-      if (!polygon.intersects(bounds) && !polygon.contains(bounds)) {
-        continue;
-      }
-      // use a copy since we will move the polygon
-      polygon = new Polygon(polygon.xpoints, polygon.ypoints, polygon.npoints);
-      polygon.translate(-bounds.x, -bounds.y);
-      graphics.setPaint(territoryPaint);
-      graphics.fillPolygon(polygon);
-      graphics.setColor(Color.BLACK);
-      graphics.drawPolygon(polygon);
-    }
-  }
-}
-
-
-class BattleDrawable extends TerritoryDrawable implements IDrawable {
-  private final String m_territoryName;
-
-  public BattleDrawable(final String territoryName) {
-    m_territoryName = territoryName;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Set<PlayerID> players = new HashSet<>();
-    for (final Unit u : data.getMap().getTerritory(m_territoryName).getUnits()) {
-      if (!TripleAUnit.get(u).getSubmerged()) {
-        players.add(u.getOwner());
-      }
-    }
-    final Territory territory = data.getMap().getTerritory(m_territoryName);
-    PlayerID attacker = null;
-    boolean draw = false;
-    for (final PlayerID p : players) {
-      if (!territory.isWater()) {
-        if (data.getRelationshipTracker().isAtWar(p, territory.getOwner())) {
-          attacker = p;
-          draw = true;
-          break;
-        } else {
-          // O(n^2), but n is usually 2, and almost always < 10
-          for (final PlayerID p2 : players) {
-            if (data.getRelationshipTracker().isAtWar(p, p2)) {
-              draw = true;
-              break;
-            }
-          }
-        }
-      } else {
-        // O(n^2), but n is usually 2, and almost always < 10
-        for (final PlayerID p2 : players) {
-          if (data.getRelationshipTracker().isAtWar(p, p2)) {
-            draw = true;
-            break;
-          }
-        }
-      }
-    }
-    if (draw) {
-      Color stripeColor;
-      if (attacker == null || territory.isWater()) {
-        stripeColor = Color.RED.brighter();
-      } else {
-        stripeColor = mapData.getPlayerColor(attacker.getName());
-      }
-      final Paint paint = new GradientPaint(0 - (float) bounds.getX(), 0 - (float) bounds.getY(),
-          new Color(stripeColor.getRed(), stripeColor.getGreen(), stripeColor.getBlue(), 120),
-          30 - (float) bounds.getX(), 50 - (float) bounds.getY(), new Color(0, 0, 0, 0), true);
-      super.draw(bounds, graphics, mapData, scaled, territory, paint);
-    }
-  }
-
-  @Override
-  public int getLevel() {
-    return BATTLE_HIGHLIGHT_LEVEL;
-  }
-}
-
-
-class LandTerritoryDrawable extends TerritoryDrawable implements IDrawable {
-  private final String m_territoryName;
-
-  public LandTerritoryDrawable(final String territoryName) {
-    m_territoryName = territoryName;
-  }
-
-  @Override
-  public void draw(final Rectangle bounds, final GameData data, final Graphics2D graphics, final MapData mapData,
-      final AffineTransform unscaled, final AffineTransform scaled) {
-    final Territory territory = data.getMap().getTerritory(m_territoryName);
-    Color territoryColor;
-    final TerritoryAttachment ta = TerritoryAttachment.get(territory);
-    if (ta != null && ta.getIsImpassible()) {
-      territoryColor = mapData.impassibleColor();
-    } else {
-      territoryColor = mapData.getPlayerColor(territory.getOwner().getName());
-    }
-    draw(bounds, graphics, mapData, scaled, territory, territoryColor);
-  }
-
-  @Override
-  public int getLevel() {
-    return POLYGONS_LEVEL;
-  }
-}
