+++ src/games/strategy/triplea/ui/screen/drawable/ReliefMapDrawable.java
@@ -1,0 +1,44 @@
+package games.strategy.triplea.ui.screen.drawable;
+
+import games.strategy.triplea.image.TileImageFactory;
+import games.strategy.triplea.ui.IUIContext;
+
+import java.awt.Image;
+
+class ReliefMapDrawable extends MapTileDrawable {
+  public ReliefMapDrawable(final int x, final int y, final IUIContext context) {
+    super(x, y, context);
+  }
+
+  @Override
+  public MapTileDrawable getUnscaledCopy() {
+    final ReliefMapDrawable copy = new ReliefMapDrawable(m_x, m_y, m_uiContext);
+    copy.m_unscaled = true;
+    return copy;
+  }
+
+  @Override
+  protected Image getImage() {
+    if (m_noImage) {
+      return null;
+    }
+    if (!TileImageFactory.getShowReliefImages()) {
+      return null;
+    }
+    Image rVal;
+    if (m_unscaled) {
+      rVal = m_uiContext.getTileImageFactory().getUnscaledUncachedReliefTile(m_x, m_y);
+    } else {
+      rVal = m_uiContext.getTileImageFactory().getReliefTile(m_x, m_y);
+    }
+    if (rVal == null) {
+      m_noImage = true;
+    }
+    return rVal;
+  }
+
+  @Override
+  public int getLevel() {
+    return RELIEF_LEVEL;
+  }
+}
